Improvement: Made the feature an independent singleton
